relicensed to CC-BY-SA-4.0
previously, this repository's work was licensed under the terms of the MIT
license. to my knowledge, there doesn't appear to be a way to opt out of
GitHub's use of your code to train Copilot. therefore, the README.md in this
repo is available under the terms of the MIT license up to and including commit
05e759. all revisions starting with this commit are licensed under the Creative
Commons Attribution-ShareAlike 4.0 International Public License.